+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## [0.4.2]
+
+- More accurate logic to identify Lightning files
+
+## [0.4.1]
+
+- Regex and Header analysis improvements
+- Tests improvements
+- Prettier code formatting
+
+## [0.4.0]
+
+- LWC Support
+
 ## [0.3.0]
 
 - Lightning Support!